<temp> FogFilterShader 수정 완료 이제 텍스쳐로 출력을 잘합니다.

+++ Project/Engine/struct.fx
@@ -166,7 +166,7 @@ struct tRayOutput
     
     int check1;
     int check2;
-    int pad[2];
+    float pad[2];
 };
